LayoutTests/fast/sandbox/ios/sandbox-mach-lookup-mail-expected.txt

Regression tests for mach lookup sandbox changes for Mail on iOS

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.hasSandboxMachLookupAccessToGlobalName("com.matthewbenedict.CyberKit.WebContent", "com.apple.lsd.open") is true
PASS internals.hasSandboxMachLookupAccessToGlobalName("com.matthewbenedict.CyberKit.WebContent", "com.apple.mobileassetd") is true
PASS internals.hasSandboxMachLookupAccessToGlobalName("com.matthewbenedict.CyberKit.WebContent", "com.apple.iconservices") is true
PASS internals.hasSandboxMachLookupAccessToGlobalName("com.matthewbenedict.CyberKit.WebContent", "com.apple.PowerManagement.control") is true
PASS internals.hasSandboxMachLookupAccessToGlobalName("com.matthewbenedict.CyberKit.WebContent", "com.apple.frontboard.systemappservices") is true
